renamed ModelStructure.py for clarity
