+++ lex/lex_app/model_utils/ModelStructure.py
@@ -1,5 +1,6 @@
 import yaml
 
+
 class ModelStructure:
     def __init__(self, path: str):
         self.path = path
@@ -9,14 +10,14 @@ class ModelStructure:
         self._load_info()
 
     def _load_info(self):
-        with open(self.path, 'r') as f:
+        with open(self.path, "r") as f:
             data = yaml.safe_load(f)
         try:
-            self.structure = data['model_structure']
+            self.structure = data["model_structure"]
         except (KeyError, TypeError):
             print("Error: Structure is not defined in the model info file")
         try:
-            self.styling = data['model_styling']
+            self.styling = data["model_styling"]
         except (KeyError, TypeError):
             print("Error: Styling is not defined in the model info file")
